Add test:monitoring script
Note networkidle0 changed to networkidle2. For some reason there is at
least a single network connection active when hitting a deployed
developer portal. Unknown if there is actually an open connection or
something strange going on inside puppeteer. Still investigating why
networkidle0 had to be changed to get this to work.