Argonaut swagger pointed to prod instead of staging
It is confusing looking at the production developer portal and seeing a
swgger doc hosted in 'staging'. It is not an option to grab the document
from the current environment. The health API swagger doc is not availble in dev.
The health API is going to be going through changes in the near future,
changing from being hosted at /argonaut to /fhir, so spending a lot of
time on this now seems wasteful. Simpliest solution is to update the
hardcoded health api swagger spec to point to production instead of
staging. This allows the spec to still be pulled down in all three
environments, with the production development portal showing that the
docuemnt is from production, which is what users expect.