Convert ServersContainer to typescript
The file uses SwaggerUI, so I had to leave some things as the any type
at this time. We could get more invested at a later date, but to be
honest, the ServersContainer is hidden with CSS anyway and unused right
now, so I don't want to spend too much time.

I did add an interface for the props specifying one level deep of what
it should have and what properties they should have.