Use React Hooks for ApiPage
There were some non-null assertions in here that eslint does not like. I
removed them but have questions on how we should handle these cases, if
they are somehow null - or if we have non-null assertions, I assume we
have a good reason. So maybe we need to do some refactoring in order to
solve those issues.