Use ReactHooks for VeteransCrisisLine
To try and clean up the component I made a custom hook for controlling
the modal. I didn't see any other modals in the code, or at least not
obvious modals, to have a reason to pull the custom Effect out into a
different file. Maybe we could create an effects/ folder in our src code
to store any custom effects we create for easy reuse?