Add PropTypes to react hooks
Note that on ApiPage, the propTypes are set to an empty object, but it
actually takes some Routing properties. This is because we want to
upgrade ApiPage to use hooks but at present we're on version 4 of
react-router-dom. In order to use the hooks, which will remove the need
for us to have any props passed to ApiPage, we need to upgrade to
version 5 of react-router-dom.