Add ModalController test
I tried to use the renderHook function from the testing library but ran
into issues with eslint claiming it was using an any type, even after I
specified the types on everyting. For the moment I went with the pattern
used on the Previous hook we made. For the most part, if we switch to
the renderHook function at a later date, the test shouldn't really
change besides how we are rendering