Add test coverage for custom hooks
The custom hooks now have better test coverage.

I added the react-hooks testing library so we can use the renderHook
function for testing our custom hooks. That said, we aren't using this
in ModalController because it requires access to the DOM for firing
'typing' events. I couldn't find a way to access them (screen doesn't
work because react-hooks uses the react-test-renderer for rendering and
works differently.