Do not throw error without error boundary defined
We don't have an error boundary component defined and it likely won't be
defined soon. Instead of throwing an error and causing the app rendering
to break, we can render a separate component in the case that the
rendering would fail (when application.result is undefined)

When we add the error boundary component, it's highly likely we can
change this back to throwing an error with a message for the error
boundary component to display. At this time we do not have the boundary
component though.