Set colors so contrast is at least 4.5:1
In order to do this, I overrided some css classes in the theme we are
using (atom-one-dark-reasonable). I created a separate scss file for the
overrides so they can be tracked separately at a later date, and if the
original theme is updated, it will be easier to track when we can remove
those overrides.