Explain why contrast checks are disabled on auth
I updated the comment explaining why we have color contrast checks
turned off on the authorization page, in the accessibility automated
tests.

Once a new version of jest-axe comes out with axe-core 4.1.0 as a
downstream dependency (and we upgrade), we should be able to turn these
checks back on. Axe-core 4.1.0 documents that it fixed this issue with
color contrast checks.