Display Not Found when category does not have api
When a user searches for an api under a category that does not have an
api of the given name in the url, we now display a 404 not found page.

For example, the health category does not include a claims api. Before
when a user would go to /explore/health/docs/claims, the claims api
under the benefits category would be displayed (confusingly also
identifying itself as a health category api). Now when a user searches,
they get a Not Found page.