Refactor code - comment/explain unreachable code
In the deactivation message of the ApiPage component, we have two
methods used to determine if an api is deprecated or deactivated. Both
of these methods return false if the api's deactivation info is
undefined. In this scenario, DecativationMessage was returning null.
Afterward in the code, the compiler does not see that api
deactivationInfo must be defined to get past that if statement that
returns null. This caused us to add additional if logic, and create
arrow functions for rendering in cases that are theoretically
unreachable.

While writing the test for this component, I was unable to reach 95%
coverage due to this. This refactor should allow me to reach 95%
coverage, and really simplify our logic.

I put a large comment explaining why the new theoretically unreachable
code exists and that, if for some reason, my logic on this isn't sound,
that that line might be a good place to start searching if
DeactivationMessages were to start disappearing from pages that should
have them