Add ApiPage test
This is the first commit for this test with a 95% coverage.

A couple notes:
- I mocked the ExplorePage, since it's an .mdx file, and it didn't seem
to be working properly in the React/Jest ecosystem.
- I mocked the ApiDocumentation component, since it was having display
issues due to not having the entire reach store configured in this test.
- I mocked the react router dom to easily change the parameters I wanted
it to have. I may removed this mock, if possible, in a later commit.
- I add an openAPI server. This may be unneeded. I'm looking into
removing this.
- I attempted to make the tests verbose. This lead to 4 indents, which
our linter does not like. We could refactor these tests to resolve the
issue.
- I used similar 'spyOn' functionality that the already existing
ApiDocumentation test uses. With this I was able to mock the existing
categories and apis without needing to mock the actual apiDefs files.