Clarify some tests
There was one test that had the wrong description.

I added convenience variables for the fake categories, so there
weren't random index values through the test. Note that if
fakeCategories are updated, then this test will need updated
appropriately. For instance, if the indexes change or if an
existing category/api is modified in a way that the test cases break

I removed the semi-global BeforeEach and have each inner describe
setting up its own mocks. Because of how long this test file is, this
helps make it more explicit within each test the scenario and that the
mocked values match that description of the test case.