Downgrade react-dev-utils to 9.1.0
With a higher version of react-dev-utils the e2e and visual regression
tests were failing. With version 9.1.0 they pass once again.

For now we're downgrading and will upgrade again at a later date. The
strategy will likely be to step one version at a time to determine if
there's something in particular that is causing the issues.

API-4498