Upgrade webpack cli and dev server to latest
Both the webpack cli and webpack dev server have been updated to their
latest versions. They appear to work fine but I need to test in another
environment. The dev build in my local seems to have an issue loading
the api specifications, but I believe this is due to the way I am
running it and not a bad package. This will be tested out and specified
in the PR

API-4493