Remove target branch from dependabot config
I had target branch in the dependabot config to try and test it out, but
I'm not sure how easy it will be to test. I think at this point it might
be best to merge this, and then on the first of next month see if the
dependabot updates come up correctly.

API-3904