Use <td> instead of <th scope="row"> for left column of the table in the "Privacy policy and terms of service" section on the Request Production Access page
This isn't actually what we want to do. <th scope="row"> is more correct – the left column is genuinely a header describing the row. However, we had an issue with the screen reader where it would announce the wrong cell's content when navigating vertically. Since this wasn't working, we went back to <td> and will attempt to fix this in a later ticket.

See here for understanding accessible table markup: https://webaim.org/techniques/tables/data